Title: 선두를 지켜라 3, Time: 5ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240710/선두를 지켜라 3/keep-the-lead-3.cpp
@@ -11,8 +11,8 @@ int main() {
     int N = 0, M = 0;
     cin >> N >> M;
 
-    unsigned int distancePerTimeA[1000]{};
-    unsigned int distancePerTimeB[1000]{};
+    unsigned long long distancePerTimeA[1000]{};
+    unsigned long long distancePerTimeB[1000]{};
 
     int timeIndex = 0;
     for (int i = 0; i < N; ++i)
